# faster replace find first of with find