--- a/include/ada/helpers.h
+++ b/include/ada/helpers.h
@@ -66,7 +66,7 @@ namespace ada::helpers {
    * Returns a host's delimiter location depending on the state of the instance.
    * Used by the host parser.
    */
-  ada_really_inline size_t get_host_delimiter_location(const ada::url& url, std::string_view& view, bool& inside_brackets) noexcept;
+  ada_really_inline size_t get_host_delimiter_location(const bool is_special, std::string_view& view, bool& inside_brackets) noexcept;
 
   /**
    * Removes leading and trailing C0 control and whitespace characters from string.