--- a/include/ada/helpers.h
+++ b/include/ada/helpers.h
@@ -63,7 +63,8 @@ namespace ada::helpers {
   ada_really_inline std::string_view substring(std::string_view input, size_t pos) noexcept;
 
   /**
-   * Returns a host's delimiter location depending on the state of the instance, and whether a colon was found.
+   * Returns a host's delimiter location depending on the state of the instance, and 
+   * whether a colon was found outside brackets.
    * Used by the host parser.
    */
   ada_really_inline std::pair<size_t,bool> get_host_delimiter_location(const bool is_special, std::string_view& view) noexcept;