This is an alternative to @miguelteixeiraa 's PR  https://github.com/ada-url/ada/pull/231

Miguel got most of the gains, but I am squeezing a few more and simplifying the code a bit...

Here are the results with the `bench` benchmark on an M2 processor (LLVM 14).

main branch
```
BasicBench_AdaURL_With_Move       2003 ns         2002 ns       350705 GHz=4.30689 cycle/byte=10.2129 cycles/url=750.182 instructions/byte=53.4653 instructions/cycle=5.23509 instructions/ns=22.547 instructions/url=3.92727k ns/url=174.182 speed=403.544M/s time/byte=2.47804ns time/url=182.023ns url/s=5.4938M/s
```

This PR:
```
BasicBench_AdaURL_With_Move       1854 ns         1854 ns       373606 GHz=4.4108 cycle/byte=10.0062 cycles/url=735 instructions/byte=50.5025 instructions/cycle=5.04712 instructions/ns=22.2619 instructions/url=3.70964k ns/url=166.636 speed=435.717M/s time/byte=2.29507ns time/url=168.583ns url/s=5.93179M/s
```

Miguel:
```
BasicBench_AdaURL_With_Move       2104 ns         2030 ns       342270 GHz=4.42335 cycle/byte=10.0347 cycles/url=737.091 instructions/byte=51.7475 instructions/cycle=5.15688 instructions/ns=22.8107 instructions/url=3.80109k ns/url=166.636 speed=397.999M/s time/byte=2.51257ns time/url=184.56ns url/s=5.41831M/s
```

The important measure for me is `instructions/url`. Or, if you prefer `instructions/byte`. It is amazing how much Miguel was able to trim out. This PR is a little bit better.

In total, it is reduction by 3% of the number of instructions per URL parsed. Wow. For such a small change...

Here are the results on an Zen 2 processor (x64) with GCC 11:

This PR:
```
BasicBench_AdaURL_With_Move       4527 ns         4527 ns       154928 GHz=3.4387 cycle/byte=19.9938 cycles/url=1.46864k instructions/byte=43.9542 instructions/cycle=2.19839 instructions/ns=7.5596 instructions/url=3.22864k ns/url=427.091 speed=178.499M/s time/byte=5.60226ns time/url=411.511ns url/s=2.43007M/s
```

Main branch:
```
BasicBench_AdaURL_With_Move       4959 ns         4959 ns       141184 GHz=3.43635 cycle/byte=22.1151 cycles/url=1.62445k instructions/byte=48.948 instructions/cycle=2.21333 instructions/ns=7.60577 instructions/url=3.59545k ns/url=472.727 speed=162.936M/s time/byte=6.13738ns time/url=450.818ns url/s=2.21819M/s
```

So a very nice gain again!!! A 9% reduction in the number of instructions.


One important code change is this...

```C++
std::string_view buffer(&*(pointer + 1));
```

to 

```C++
std::string_view buffer = new_host.substr(location+1);
```

I think that the latter is much nicer.

There is also another place where we had this crazy if clause that was entirely unnecessary because we it was rechecking the result of get_host_delimiter_location.

The updated code also uses SWAR to find the next matching character.