@lemire 2023-02-24
Ah. I need to add a check for endianness... of course...