@lemire 2023-02-24
Damn it. The big endian stuff is still broken?