@lemire 2023-02-25
@anonrig I should have fixed the big-endian support now. Please review again.

> What are the new benchmark results? Can you update the description?

It should be current.